Moved api to new directory contract
